pyproject.toml
[tool.poetry]
name = "ss14-status-bot"
version = "0.1.0"
description = ""
authors = ["todotom <tomasdarioam@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 140
